Netzwerk Interfaces  überarbeitet und udp send implementiert
git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@906 31957799-4632-4615-9793-72db6d48bcac

+++ user/user_tools/userlib.hpp
@@ -6,8 +6,8 @@
 extern "C" {
 #include "userlib.h"
 
-void* operator new      (long unsigned int size);
-void* operator new[]    (long unsigned int size);
+void* operator new      (size_t size);
+void* operator new[]    (size_t size);
 void  operator delete   (void* ptr);
 void  operator delete[] (void* ptr);
 }
