+++ user/user_tools/userlib_cpp.cpp
@@ -6,11 +6,11 @@
 #include "userlib.hpp"
 #include "stdlib.h"
 
-void* operator new(long unsigned int size)
+void* operator new(size_t size)
 {
     return malloc(size);
 }
-void* operator new[](long unsigned int size)
+void* operator new[](size_t size)
 {
     return malloc(size);
 }
